# docs/swagger.yaml
basePath: /
definitions:
  service.CreateSessionRequest:
    type: object
  service.CreateWorkoutLogRequest:
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend API for the Heracle personal training & fitness platform.
  title: Heracle Fitness API
  version: "1.0"
paths:
...
  /workout/exercises:
    get:
      description: Returns the full exercise library with image URLs
      produces:
      - application/json
      responses:
        "200":
          description: Exercise list
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all exercises
      tags:
      - Workout
  /workout/log:
    post:
      consumes:
      - application/json
      description: Logs a completed workout for the authenticated user
      parameters:
      - description: Workout log data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.CreateWorkoutLogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Created workout log
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a workout log
      tags:
      - Workout
  /workout/log/{id}:
    delete:
      description: Deletes a workout log by ID
      parameters:
      - description: Workout log ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Deleted
        "404":
          description: Log not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a workout log
      tags:
      - Workout
    get:
      description: Returns a specific workout log by ID
      parameters:
      - description: Workout log ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Workout log
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Log not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a workout log
      tags:
      - Workout
    patch:
      consumes:
      - application/json
      description: Updates an existing workout log by ID
      parameters:
      - description: Workout log ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Updated workout log
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Log not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a workout log
      tags:
      - Workout
  /workout/logs:
    get:
      description: Returns all workout logs for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Workout logs
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all workout logs
      tags:
      - Workout
  /workout/my-sessions:
    get:
      description: Returns all custom workout sessions for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: User sessions
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user's workout sessions
      tags:
      - Workout
  /workout/preferences:
    get:
      description: Returns workout preferences for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Workout preferences
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get workout preferences
      tags:
      - Workout
    post:
      consumes:
      - application/json
      description: Saves or updates workout preferences for the authenticated user
      parameters:
      - description: Workout preferences
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Updated preferences
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Save workout preferences
      tags:
      - Workout
  /workout/session:
    post:
      consumes:
      - application/json
      description: Creates a new custom workout session for the authenticated user
      parameters:
      - description: Session data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.CreateSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Created session
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a workout session
      tags:
      - Workout
  /workout/session/{id}:
    delete:
      description: Deletes a workout session by ID
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Deleted
        "404":
          description: Session not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a workout session
      tags:
      - Workout
    get:
      description: Returns a specific workout session by ID
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Session details
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Session not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a workout session
      tags:
      - Workout
    patch:
      consumes:
      - application/json
      description: Updates an existing workout session by ID
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Updated session
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Session not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a workout session
      tags:
      - Workout
  /workout/sessions:
    get:
      description: Returns hardcoded sample workout sessions
      produces:
      - application/json
      responses:
        "200":
          description: Static sessions
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: Get static workout sessions
      tags:
      - Workout
  /workout/today:
    get:
      description: Returns a workout suggestion based on user preferences
      produces:
      - application/json
      responses:
        "200":
          description: Workout suggestion
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get today's workout suggestion
      tags:
      - Workout
  /workout/trainer/log-review/{logId}:
    patch:
      consumes:
      - application/json
      description: Allows a trainer to add a review to their client's workout log
      parameters:
      - description: Workout log ID
        in: path
        name: logId
        required: true
        type: integer
      - description: Review data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Updated workout log with review
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Not authorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Log not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 'Trainer: add review to workout log'
      tags:
      - Workout
  /workout/trainer/session/{clientId}:
    post:
      consumes:
      - application/json
      description: Allows a trainer to create a workout session for their assigned
        client
      parameters:
      - description: Client user ID
        in: path
        name: clientId
        required: true
        type: string
      - description: Session data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.CreateSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Created session
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Not authorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 'Trainer: create client session'
      tags:
      - Workout
  /workout/trainer/session/{clientId}/{sessionId}:
    delete:
      description: Allows a trainer to delete a workout session for their assigned
        client
      parameters:
      - description: Client user ID
        in: path
        name: clientId
        required: true
        type: string
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: integer
      responses:
        "200":
          description: Deleted
        "403":
          description: Not authorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Session not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 'Trainer: delete client session'
      tags:
      - Workout
    patch:
      consumes:
      - application/json
      description: Allows a trainer to update a workout session for their assigned
        client
      parameters:
      - description: Client user ID
        in: path
        name: clientId
        required: true
        type: string
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Updated session
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Not authorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Session not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 'Trainer: update client session'
      tags:
      - Workout
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"